src/code_runner_backend/test.py
import os
from dotenv import load_dotenv
from openai import AsyncOpenAI

# Load environment variables
load_dotenv()

# Read Azure OpenAI configuration
AZURE_OPENAI_API_KEY = os.getenv("AZURE_OPENAI_API_KEY")
AZURE_OPENAI_ENDPOINT = os.getenv("AZURE_OPENAI_ENDPOINT")
AZURE_OPENAI_DEPLOYMENT = os.getenv("AZURE_OPENAI_DEPLOYMENT")

client = AsyncOpenAI(
    api_key=AZURE_OPENAI_API_KEY,
    base_url=AZURE_OPENAI_ENDPOINT
)

async def test_azure_openai():
    print("⚙️ Requesting Azure OpenAI...")
    
    response = await client.chat.completions.create(
        model=AZURE_OPENAI_DEPLOYMENT,
        messages=[
            {"role": "user", "content": "Hello, please write a Python program to print hello world"}
        ]
    )
    
    print("✅ GPT Response:\n")
    print(response.choices[0].message.content)

if __name__ == "__main__":
    asyncio.run(test_azure_openai())
